pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bid_scoring"
version = "0.1.0"
description = "Bid Scoring System"
requires-python = ">=3.11"
dependencies = [
    "fastmcp",
    "minio",
    "openai",
    "pgvector",
    "psycopg[binary]",
    "psycopg-pool>=3.3.0",
    "pymupdf>=1.23.0",
    "python-dotenv",
    "pyyaml",
    "requests",
    "tenacity",
    "tiktoken",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["bid_scoring*", "mcp_servers*", "mineru*"]

[project.scripts]
bid-scoring-retrieval = "mcp_servers.retrieval_server:main"
bid-pipeline = "bid_scoring.pipeline.interfaces.cli:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23.0",
    "ruff",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "unit: marks tests as pure unit tests",
    "integration: marks tests requiring db/filesystem integration",
    "e2e: marks tests as end-to-end (slow, requires external services)",
]
